requirements.txt
fastapi
uvicorn
requests
google-genai
arize-phoenix
arize-phoenix-evals
opentelemetry-instrumentation-fastapi
python-dotenv
pydantic-settings
cachetools
scipy